requirements.txt
Flask==2.0.2
requests==2.26.0
beautifulsoup4==4.10.0
python-dotenv==0.19.2
selenium==4.11.2
huggingface_hub==0.16.4